+++ packages/orq-rc/src/models/components/reasoningpart.ts
@@ -40,7 +40,7 @@ export const ReasoningPart$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("reasoning_01ky56kdh7hxbzg09z7jztcwag"),
+  _id: z.string().default("reasoning_01ky5zbr0ja4w046xjfa1q7ytr"),
   metadata: z.record(z.any()).optional(),
   kind: z.literal("reasoning"),
   reasoning: z.string(),