+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -4054,7 +4054,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KY56KDQ2WZ7Q4D5XMT376F1A"),
+  id: z.string().default("01KY5ZBR5DRQZCWZ7G1PT64NR3"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),