+++ packages/orq-rc/RELEASES.md
@@ -11099,3 +11099,13 @@ Based on:
 - [typescript v4.12.0-rc.56] packages/orq-rc
 ### Releases
 - [NPM v4.12.0-rc.56] https://www.npmjs.com/package/@orq-ai/node/v/4.12.0-rc.56 - packages/orq-rc
+
+## 2026-07-22 22:35:46
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.790.3 (2.918.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.13.0-rc.0] packages/orq-rc
+### Releases
+- [NPM v4.13.0-rc.0] https://www.npmjs.com/package/@orq-ai/node/v/4.13.0-rc.0 - packages/orq-rc